third_party/blink/web_tests/platform/linux-chrome/external/wpt/loading/early-hints/preconnect-in-early-hints.h2.window-expected.txt

This is a testharness.js-based test.
[FAIL] Preconnect in early hints.
  assert_equals: Connection establishment should not take time for a resource from a preconnected origin. expected 173.5 but got 169.80000000074506
Harness: the test ran to completion.
